Refactor functional app setup
This commit updates the application with the latest API changes in
Spring Framework and uses the latest Spring Data version.

This also delegates the server setup to the experimental boot starter
and only create the HttpHandler, which will be picked up by the embedded
server itself.
