+++ src/main/java/com/example/functional/WebFunctionPlaygroundApplication.java
@@ -5,23 +5,24 @@ import com.example.functional.domain.User;
 import com.example.functional.domain.UserRepository;
 import com.example.functional.web.UserHandler;
 import reactor.core.publisher.Mono;
-import reactor.ipc.netty.http.HttpServer;
 
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.data.mongo.MongoDataAutoConfiguration;
 import org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration;
+import org.springframework.boot.autoconfigure.reactiveweb.ReactiveWebAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.http.server.reactive.HttpHandler;
-import org.springframework.http.server.reactive.ReactorHttpHandlerAdapter;
-import org.springframework.web.reactive.function.RouterFunction;
+import org.springframework.web.reactive.function.server.RouterFunction;
 
-import static org.springframework.web.reactive.function.RequestPredicates.*;
-import static org.springframework.web.reactive.function.RouterFunctions.*;
+import static org.springframework.web.reactive.function.server.RequestPredicates.GET;
+import static org.springframework.web.reactive.function.server.RequestPredicates.POST;
+import static org.springframework.web.reactive.function.server.RouterFunctions.route;
+import static org.springframework.web.reactive.function.server.RouterFunctions.toHttpHandler;
 
 @SpringBootApplication(
-		exclude = {MongoAutoConfiguration.class, MongoDataAutoConfiguration.class})
+		exclude = {MongoAutoConfiguration.class, MongoDataAutoConfiguration.class, ReactiveWebAutoConfiguration.class})
 public class WebFunctionPlaygroundApplication {
 
 	public static void main(String[] args) {
@@ -44,11 +45,18 @@ public class WebFunctionPlaygroundApplication {
 	}
 
 	@Bean
+	HttpHandler httpHandler(RouterFunction<?> router) {
+		return toHttpHandler(router);
+	}
+
+	/*
+	@Bean
 	HttpServer nettyServer(RouterFunction<?> router) {
 		HttpHandler handler = toHttpHandler(router);
 		HttpServer httpServer = HttpServer.create(7080);
 		httpServer.start(new ReactorHttpHandlerAdapter(handler));
 		return httpServer;
 	}
+	*/
 
 }
