+++ src/main/java/com/example/functional/web/UserHandler.java
@@ -7,8 +7,8 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import org.springframework.stereotype.Component;
-import org.springframework.web.reactive.function.ServerRequest;
-import org.springframework.web.reactive.function.ServerResponse;
+import org.springframework.web.reactive.function.server.ServerRequest;
+import org.springframework.web.reactive.function.server.ServerResponse;
 
 @Component
 public class UserHandler {
@@ -19,18 +19,18 @@ public class UserHandler {
 		this.repository = repository;
 	}
 
-	public ServerResponse<Publisher<User>> listPeople(ServerRequest request) {
+	public Mono<ServerResponse> listPeople(ServerRequest request) {
 		Flux<User> people = this.repository.findAll();
 		return ServerResponse.ok().body(people, User.class);
 	}
 
-	public ServerResponse<Publisher<User>> getUser(ServerRequest request) {
+	public Mono<ServerResponse> getUser(ServerRequest request) {
 		String personId = request.pathVariable("id");
 		Mono<User> person = this.repository.findOne(personId);
 		return ServerResponse.ok().body(person, User.class);
 	}
 
-	public ServerResponse<Mono<Void>> createUser(ServerRequest request) {
+	public Mono<ServerResponse> createUser(ServerRequest request) {
 		Mono<User> user = request.bodyToMono(User.class);
 		return ServerResponse.ok().build(this.repository.save(user).then());
 	}
